ref: split processing terminal outcomes
