+++ src-tauri/src/processing/AGENTS.md
@@ -10,7 +10,7 @@
   `ProgressEmitter`.
 
 ## Private Cluster
-- Files: `../processing.rs`, `plan.rs`, `run.rs`, `terminal_outcomes.rs`,
+- Files: `../processing.rs`, `plan.rs`, `run.rs`, `terminal_outcomes/`,
   `lifecycle.rs`, `context/`, `job_registry/`, `progress/`,
   `preview_config.rs`, `session.rs`, `contract_tests.rs`.
 - The cluster owns preflight planning, execution-plan preparation, runner
@@ -21,7 +21,9 @@
 ## Allowed Agent Edits Without Escalation
 - Change planner or runner internals when `bun scripts/proof/runner.ts focus rust contract` stays green.
 - Keep preflight side-effect-free; execution may create output dirs only after review enforcement.
-- Keep runner responsibilities to encoder/toolchain validation, events, job registration, scheduler dispatch, audio execution requests through `crate::audio`, and terminal normalization.
+- Keep runner responsibilities to encoder/toolchain validation, job
+  registration, scheduler dispatch, audio execution requests through
+  `crate::audio`, and handoff to terminal outcome helpers.
 - Keep metadata save reporting lifecycle truth through this strip while leaving
   metadata write policy inside metadata-owned APIs.
